requirements.txt
fastapi==0.109.2
pydantic==1.10.13
mangum==0.17.0
python-dotenv==1.0.0
uvicorn==0.27.1
httpx
langchain-community>=0.2.0
langchain-openai
openai
langgraph>=0.0.10
langchain-core>=0.1.17